requirements.txt
flask
langchain
langchain-community
langchain-groq
python-dotenv
sqlalchemy
mysql-connector-python